RAG_IC&MS/pyproject.toml
[tool.poetry]
name = "customer-support-rag"
version = "0.1.0"
description = "RAG-based customer support system for generating draft responses"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = { version = "^0.34", extras = ["standard"] }
openai = "^1.60"
pinecone-client = "^5.0"
python-dotenv = "^1.0"
pydantic-settings = "^2.7"
tiktoken = "^0.8"
PyMuPDF = "^1.25"
pdfplumber = "^0.11"
pytesseract = "^0.3"
python-docx = "^1.1"
extract-msg = "^0.50"
pandas = "^2.2"
langdetect = "^1.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.25"
httpx = "^0.28"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
